Use CGFloat in roundToStep

+++ TurtleBezierPath demo/TurtleBezierPath demo/RoundedUISlider.m
@@ -28,7 +28,7 @@
 
 #pragma mark - Maths
 
-static inline float roundToStep( float number, float step )
+static inline CGFloat roundToStep( CGFloat number, CGFloat step )
 {
     return ( step != 0.0f ) ? floorf( number / step ) * step : number;
 }
